Refactor the likelihood into separate subclasses
